app/requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
prometheus-client==0.11.0
boto3==1.18.65
